backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-google-genai==0.0.8
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
fastapi-cors==0.0.6
langchain-community==0.0.10
faiss-cpu==1.7.4
